Splitting code into separate files for different functionalities and refacored them
